# springdoc-openapi/swagger-generator/src/main/java/com/alan/mall/swagger/swaggerui/bookshop.json
{
  "swagger": "2.0",
  "info": {
    "description": "RESTful Swagger API 生成器",
    "version": "1.0.0",
    "title": "Swagger 脚本生成器",
    "termsOfService": "admin",
    "contact": {
      "email": "admin"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "petstore.swagger.io",
  "basePath": "/v2",
...
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/test1/{id}": {
      "delete": {
        "tags": [
          "Test1Controller"
        ],
        "summary": "根据ID删除Test1",
        "description": "根据ID删除Test1方法详情",
        "operationId": "deleteUsingDELETE",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "根据ID删除Test1方法详情",
            "required": true
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "根据ID删除Test1"
            ,
            "schema": {
              "$ref": "#/definitions/RespResult"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Test1Controller"
        ],
        "summary": "根据ID修改Test1",
        "description": "根据ID修改Test1方法详情",
        "operationId": "updateUsingPUT",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "根据ID修改Test1方法详情",
            "required": true
          },
          {
            "in": "body",
            "name": "Test1",
            "description": "传入Test1的JSON对象",
            "required": false
              ,
              "schema": {
                "$ref": "#/definitions/Test1"
              }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "根据ID修改Test1"
            ,
            "schema": {
              "$ref": "#/definitions/RespResult"
            }
          }
        }
      },
      "get": {
        "tags": [
          "Test1Controller"
        ],
        "summary": "根据ID查询Test1",
        "description": "根据ID查询Test1方法详情",
        "operationId": "findByIdUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "根据ID修改Test1方法详情",
            "required": true
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "根据ID修改Test1"
            ,
            "schema": {
              "$ref": "#/definitions/RespResult«Test1»"
            }
          }
        }
      }
    },
...
    "/test2": {
      "get": {
        "tags": [
          "Test2Controller"
        ],
        "summary": "查询所有Test2",
        "description": "查询所有Test2方法详情",
        "operationId": "findAllUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "查询所有Test2"
            ,
            "schema": {
              "$ref": "#/definitions/RespResult«List«Test2»»"
            }
          }
        }
      },
      "post": {
        "tags": [
          "Test2Controller"
        ],
        "summary": "添加Test2",
        "description": "添加Test2方法详情",
        "operationId": "addUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "添加Test2方法详情",
            "required": true
              ,
              "schema": {
                "$ref": "#/definitions/Test2"
              }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "添加Test2"
            ,
            "schema": {
              "$ref": "#/definitions/RespResult"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "RespResult": {
      "type": "object",
      "required": [
        "code",
        "data",
        "message"
      ],
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32",
          "description": "返回状态码,20000:成功,20001:失败,20002:用户名或密码错误,20003:权限不足,20004:远程调用失败,20005:重复操作,20006:没有对应的数据"
        },
        "data": {
          "type": "object",
          "description": "逻辑数据"
        },
        "message": {
          "type": "string",
          "description": "提示信息"
        }
      },
      "description": "RespResult"
    },
...
    "RespResult«Test1»": {
      "type": "object",
      "required": [
        "code",
        "data",
        "message"
      ],
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32",
          "description": "返回状态码,20000:成功,50000:失败"
        },
        "data": {
          "description": "逻辑数据",
          "$ref": "#/definitions/Test1"
        },
        "message": {
          "type": "string",
          "description": "提示信息"
        }
      },
      "description": "RespResult"
    },
    "Test1": {
      "type": "object",
      "properties": {
        "userid": {
          "type": "integer",
          "format": "int32",
          "description": ""
        },
        "name": {
          "type": "string",
          "description": ""
        },
        "passowrd": {
          "type": "string",
          "description": ""
        },
        "phone": {
          "type": "string",
          "description": ""
        }
      },
      "description": "Test1"
    },
    "RespResult«List«Test2»»": {
      "type": "object",
      "required": [
        "code",
        "data",
        "message"
      ],
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32",
          "description": "返回状态码,20000:成功,20001:失败,20002:用户名或密码错误,20003:权限不足,20004:远程调用失败,20005:重复操作,20006:没有对应的数据"
        },
        "data": {
          "type": "array",
          "description": "逻辑数据",
          "items": {
            "$ref": "#/definitions/Test2"
          }
        },
        "message": {
          "type": "string",
          "description": "提示信息"
        }
      },
      "description": "RespResult"
    },
...
    "Test2": {
      "type": "object",
      "properties": {
        "openid": {
          "type": "integer",
          "format": "int32",
          "description": ""
        },
        "sex": {
          "type": "string",
          "description": ""
        },
        "address": {
          "type": "string",
          "description": ""
        },
        "age": {
          "type": "integer",
          "format": "int32",
          "description": ""
        },
        "testid": {
          "type": "integer",
          "format": "int32",
          "description": ""
        }
      },
      "description": "Test2"
    },
...
  },
...
}